+++ packages/cn-international-express-sdk/src/index.ts
@@ -5,3 +5,5 @@ export * from '@cn-international-express-sdk/yanwen';
 export * from '@cn-international-express-sdk/cainiao';
 export * from '@cn-international-express-sdk/jiehang';
 export * from '@cn-international-express-sdk/one7feia';
+export * from '@cn-international-express-sdk/kts';
+export * from '@cn-international-express-sdk/cne';