+++ packages/fpx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cn-international-express-sdk/fpx
 
+## 1.0.0
+
+### Major Changes
+
+- include kts-sdk fpx-sdk
+
 ## 0.2.0
 
 ### Minor Changes